fix list assinging

+++ src/user-list/user-list.service.ts
@@ -1,16 +1,29 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { UserListDto } from './dto/user-list.dto';
+import { User } from '@prisma/client';
 
 @Injectable()
 export class UserListService {
   constructor(private readonly prisma: PrismaService) {}
 
   async createRelation(dto: UserListDto) {
+    let user: User;
+
+    if (!dto.userId) {
+      user = await this.prisma.user.findUnique({
+        where: {
+          email: dto.userEmail,
+        },
+      });
+
+      if (!user) throw new NotFoundException('User not found');
+    }
+
     await this.prisma.userList.create({
       data: {
         listId: dto.listId,
-        userId: dto.userId,
+        userId: dto.userId ?? user.id,
       },
     });
 
@@ -18,11 +31,23 @@ export class UserListService {
   }
 
   async deleteRelation(dto: UserListDto) {
+    let user: User;
+
+    if (!dto.userId) {
+      user = await this.prisma.user.findUnique({
+        where: {
+          email: dto.userEmail,
+        },
+      });
+
+      if (!user) throw new NotFoundException('User not found');
+    }
+
     await this.prisma.userList.delete({
       where: {
         userId_listId: {
           listId: dto.listId,
-          userId: dto.userId,
+          userId: dto.userId ?? user.id,
         },
       },
     });
